Mother fucking big push - Before big integration

+++ MapTargetFull/src/com/example/maptargetfull/DrawSample.java
@@ -93,8 +93,8 @@ public class DrawSample extends View {
 
 			for (Friend iterable : GlobalParams.getInstance().getFriends()) {
 
-				canvas.drawBitmap(mBitmap, iterable.getWidth(),
-						iterable.getHeight(), mPaint);
+				canvas.drawBitmap(mBitmap, ((Double)iterable.getWidth()).intValue(),
+						((Double)iterable.getHeight()).intValue(), mPaint);
 				// iterable.setWidth(iterable.getWidth() + 1);
 
 			}
